Create vdbc_acura_ilx_2016_can.dbc
The file now follow the Vector dbc format. 
Byte order is now correctly set in the big endian format (motorola). 
IDs are now in decimal. 
Added the node list that was missing.
Added the NEO node (which is follow some of the ADAS behavior).
Added the INTERCEPTOR node. 
Added the EBCM node to use it as a receiver for the BRAKE_COMMAND.
Mapped all signals to a receiver node. 
Remapped GAS_SENSOR and GAS_COMMAND messages from ADAS to INTERCEPTOR.
There's only one small problem left : the LKAS_HUD2 layout need to be fixed. The mapping for the signal SET_ME_X48 doesn't work. This need to be removed or remap.