Create vdbc_acura_ilx_2016_nidec.dbc
The file now follow the Vector dbc format. 
Byte order is now correctly set in the big endian format (motorola). 
IDs are now in decimal. 
Added the node list that was missing.
Added the NEO node (which is follow some of the ADAS behavior).
Mapped all signals to a receiver node.