Reduce default gain values
While testing logic to auto-tune the angle and rate gain values, I found that the rate gain value can be **much smaller** without losing its effectiveness.  My Honda Accord performed well with a rate gain below 0.01, so it seems that 0.2 is not as conservative as I first thought.  

I've also reduced the default angle gain, since 1.0 should perform nearly the same with the proposed logic as it does with standard logic.