Removed steer_override as a condition for gain adjustment
It originally seemed that steer_override would be a condition that should block auto-adjust of the feed-forward gain values.  However, after testing on other vehicles, it was clear that OP's torque requests alone can trigger false positives for steering overrides (specifically Toyota's).  This resulted in the rate-based feed-forward gain value to go off-scale positive, which caused extremely poor steering performance in a short period of time.