RAV4H Corolla fix for steering unavailable
When the steering angle or rate is above 100 then the steering fault occurs.
To stop this from happening treat this as a criteria to switch off steering and make sure not to trip panda safety limits by not holding to the torque limit. i.e. Setting apply steer from 1500 to 0.