2019 Camry LE FPv2  (not RAV4)
My Comma 2 is giving "car not recognized", so I am submitting the fingerprint.
I am just raw copying the command I entered in Workbench and the data I received.
The car was turned on and off 3 times and the command was run 3 times. 
I pasted the 3rd data set above.
Obviously it is not a RAV4 as shown in the dump.
Noob.