Necessary changes for the LEXUS IS Hybrid were never merged in; Remove entry in README.md
https://github.com/commaai/openpilot/pull/536 was never merged in.

`master` currently does not contain anything remotely equivalent to the changes in that PR.

Considering the signal changes in that PR for the IS Hybrid, there's no way the IS Hybrid will work with current `master`. 

I have found no evidence of an IS 300H being used successfully with openpilot in Discord. In fact, the one user who was struggling on Discord was confused by @eFiniLan 's Youtube videos of the system working. There does seem to be support in the dragonpilot fork: 

https://github.com/dragonpilot-community/dragonpilot/blob/ce48b036784dc817d141f9fbd2f887f044843b2c/selfdrive/car/toyota/carstate.py#L174

I'll be working with a user to refurbish @eFiniLan 's changes and hopefully this entry in the README can be re-added. 😄