Enable Tethering when On Road
When my car is in motion I want to enable the wifi tethering, so I can debug and connect to comma
on the road.  When I am parking back at home I want it disabled so that my C2 uploads over wifi all the files.
This toggle detects when car is on road and enables wifi tethering, when state changes to offroad it disables it.