Use Wide Road Camera as Light Sensor for Screen Brightness (#23610)
Using the narrow camera as a light sensor meant that the
reflection of our own headlights, tail lights, and head lights
of opposing vehicles dominated the sensor value at night
even though the broader overall scene is very dark.

This has very little effect during daylight hours.  The change
affects nightime driving the most, resulting in a dimmer
screen when it is dark.

I think this makes more sense as the wide angle camera's
field of view is much closer to the human eye's field of
view than the narrow road camera.