FPv2: Allow excluded types as last resort
Fixes an issue that causes the car to be unrecognized
when only 1 essential ECU firmware is detected but
there are at least 3 matching excluded ECUs.

This commit refactors match_fw_to_car_fuzzy
to allow excluded types as last resort.
This is done by keeping track of the excluded
and included matches rather than leaving out
excluded types. If there is only one included
match, it then checks that the excluded match
candidate is the same as the included match candidate.